src/requirements.txt
docxtpl
python-docx
matplotlib
lxml
numpy